Successful(ish) spiral fit - doesn't reproduce previous fits with same parameters
Not convinced sort2 is doing density sort properly, may have to restructure code

	new file:   spiralfind.params
	modified:   src/classify_by_eigenvalue.f90
	modified:   src/extract_spiral_elements.f90
	modified:   src/read_eigenvalue_file.f90
	modified:   src/sort2.f90
	modified:   src/spiraldata_module.f90
	modified:   src/spiralfind.f90
	modified:   tache.params

+++ src/classify_by_eigenvalue.f90
@@ -10,7 +10,7 @@ implicit none
 
 integer,intent(inout) :: classification
 real, intent(in) :: threshold
-real,dimension(4), intent(inout) :: classnum
+integer,dimension(4), intent(inout) :: classnum
 real,dimension(3) :: values
 
 integer :: k
